requirements.txt
# Core Framework
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0

# HTTP and OAuth
requests>=2.28.0
requests-oauthlib>=1.3.0

# File System and Async I/O
watchdog>=3.0.0
aiofiles>=23.0.0
asyncio-mqtt==0.16.1

# AI and Image Processing
google-generativeai>=0.3.0
Pillow>=9.0.0

# Monitoring and Logging
structlog>=23.0.0
psutil>=5.9.0

# Web Features
python-multipart>=0.0.6
websockets>=11.0.0 